feat: Add create channel mutation and modal component
This commit adds a new mutation called "create" to the channels API in the "convex/channels.ts" file. The "create" mutation allows creating a new channel with a specified name and workspace ID. Additionally, a new modal component called "CreateChannelModal" is added to the "src/features/channels/components" directory. This modal provides a user interface for creating a new channel. These changes enhance the functionality of the application by enabling the creation of channels within workspaces.