feat: Generate new join code for workspace
This commit adds a new mutation called "newJoinCode" to the workspaces API in the "convex/workspaces.ts" file. The "newJoinCode" mutation generates a new join code for a workspace, which can be used to invite people to join the workspace. This change enhances the functionality of the application by allowing workspace admins to easily generate new join codes for their workspaces.